Update Stock.java
Stock.price changed to double; more realistic design

+++ src/ex014/Stock.java
@@ -6,7 +6,7 @@ You are creating a Stock class.
 1. Instance Variables
 	tickerSymbol (String)
 	companyName (String)
-	price (int)
+	price (double)
 	percentChange (double)
 	totalShares(int)
 	marketCap(long)
